Utilization Chart Date Range Entry Change and Auto Updateing Utilization Gauge
Changed the method of selecting past parking lot utilization to view in the line chart from day, week, and 3 weeks. Users can now enter a custom number of past days for which they want to view lot utilization in the chart.

Changed the current lot utilization gauge so that it syncronizes itself with the database every second. Also, changed the color scheme of the utilization gauge to green, yellow, and red.

Co-Authored-By: Joey Kilgore <joey.a.kilgore@gmail.com>